dll_0B_func09: type texCount u8 so byte-mask lands in saved reg (98.85->99.01)
The texCount = (u8)(*(u16*)(tex+0x10)>>8) assignment compiled to clrlwi into
scratch + mr into the saved reg; retyping the local u8 lets the mask land
directly in the saved reg, dropping the spurious mr copy.

+++ src/main/dll/dll_000B_dll0b.c
@@ -2229,7 +2229,7 @@ int dll_0B_func09(void* a0, int a1, int a2, u8 a3, void* a4)
     void* buf1;
     u8 aligned;
     void* tex;
-    int texCount;
+    u8 texCount;
     int n131p1;
     int n131;
     f32 dirX;
